+++ integration_tests/examples/demo/inputs.rs
