+++ integration_tests/examples/demo/inputs/rtp.rs
@@ -10,12 +10,14 @@ use serde_json::json;
 use strum::{Display, EnumIter, IntoEnumIterator};
 use tracing::error;
 
-use crate::smelter_state::{
-    get_free_port,
+use crate::{
     inputs::{AudioDecoder, InputHandler, VideoDecoder},
     players::InputPlayerOptions,
-    TransportProtocol, IP,
+    smelter_state::TransportProtocol,
+    IP,
 };
+
+use crate::generators::get_free_port;
 
 #[derive(Debug, Display, EnumIter, Clone)]
 pub enum RtpRegisterOptions {
